requirements.txt
ruff==0.11.7
python-dotenv==1.0.1
aiosqlite==0.22.1
aiogram==3.24.0
apscheduler==3.11.2